+++ src/layout-text.ts
@@ -1509,6 +1509,9 @@ class ContiguousBoxBuilder {
 
   open(inline: Inline, linebox: Linebox, naturalStart: boolean, start: number, blockOffset: number) {
     const box = this.opened.get(inline);
+
+    start = Math.round(start);
+
     if (box) {
       box.end = start;
     } else {
@@ -1526,6 +1529,9 @@ class ContiguousBoxBuilder {
 
   close(inline: Inline, naturalEnd: boolean, end: number) {
     const box = this.opened.get(inline);
+
+    end = Math.round(end);
+
     if (box) {
       const list = this.closed.get(inline);
       box.end = end;